refactor!: move sqlparser conversion to its own mod
