+++ crates/proof-of-sql/src/sql/logical_plans/mod.rs
@@ -2,6 +2,7 @@
 mod error;
 pub use error::LogicalPlanError;
 mod expr;
-pub use expr::Expr;
+pub use expr::{BinaryOperator, Expr};
 mod plan;
 pub use plan::LogicalPlan;
+mod sqlparser;
